refactor(components.validator): move to telecomUniverseComponents
