--- client/components/validator/validator.constant.js
@@ -1,3 +1,0 @@
-angular.module('managerApp')
-  .constant('punycode', punycode)
-  .constant('validator', validator);
